test - rename test class
